update files and fix ft_printf.c

+++ srcs/fnc_src.h
@@ -6,7 +6,7 @@
 /*   By: aerrajiy <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/23 17:52:26 by aerrajiy          #+#    #+#             */
-/*   Updated: 2022/10/23 20:53:23 by aerrajiy         ###   ########.fr       */
+/*   Updated: 2022/10/24 02:02:03 by aerrajiy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int ft_putnbr(int fd, int n);
 int ft_putunsigned(int fd, unsigned int n);
 int ft_putchar(int fd, char c);
 int ft_putstr(int fd, char *str);
-int count_digits(int n);
+int count_digits(long n);
 size_t ft_strlen(const char *s);
 
 #endif
